Age limit

+++ app/org/sagebionetworks/bridge/services/StormPathUserAdminService.java
@@ -172,9 +172,10 @@ public class StormPathUserAdminService implements UserAdminService {
         // This user may have never consented to research. Ignore if that's the case.
         for (String trackerId : study.getTrackers()) {
             Tracker tracker = studyService.getTrackerByIdentifier(trackerId);
-            
-            HealthDataKey key = new HealthDataKey(study, tracker, user);
-            healthDataService.deleteHealthDataRecords(key);
+            if (tracker != null) {
+                HealthDataKey key = new HealthDataKey(study, tracker, user);
+                healthDataService.deleteHealthDataRecords(key);
+            }
         }
     }
